reorganise messaging folders
