+++ packages/twenty-server/src/workspace/messaging/commands/crons/fetch-all-workspaces-messages.job.ts
@@ -11,7 +11,7 @@ import {
 } from 'src/core/feature-flag/feature-flag.entity';
 import { MessageQueue } from 'src/integrations/message-queue/message-queue.constants';
 import { MessageQueueService } from 'src/integrations/message-queue/services/message-queue.service';
-import { ConnectedAccountService } from 'src/workspace/messaging/connected-account/connected-account.service';
+import { ConnectedAccountService } from 'src/workspace/messaging/repositories/connected-account/connected-account.service';
 import {
   GmailPartialSyncJobData,
   GmailPartialSyncJob,
