+++ packages/twenty-server/src/workspace/messaging/repositories/connected-account/connected-account.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { ConnectedAccountService } from 'src/workspace/messaging/connected-account/connected-account.service';
+import { ConnectedAccountService } from 'src/workspace/messaging/repositories/connected-account/connected-account.service';
 import { WorkspaceDataSourceModule } from 'src/workspace/workspace-datasource/workspace-datasource.module';
 
 @Module({
