+++ packages/twenty-server/src/workspace/messaging/repositories/message-channel-message-association/message-channel-message-assocation.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { MessageChannelMessageAssociationService } from 'src/workspace/messaging/message-channel-message-association/message-channel-message-association.service';
+import { MessageChannelMessageAssociationService } from 'src/workspace/messaging/repositories/message-channel-message-association/message-channel-message-association.service';
 import { WorkspaceDataSourceModule } from 'src/workspace/workspace-datasource/workspace-datasource.module';
 
 @Module({
