+++ packages/twenty-server/src/workspace/messaging/repositories/message-channel/message-channel.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { MessageChannelService } from 'src/workspace/messaging/message-channel/message-channel.service';
+import { MessageChannelService } from 'src/workspace/messaging/repositories/message-channel/message-channel.service';
 import { WorkspaceDataSourceModule } from 'src/workspace/workspace-datasource/workspace-datasource.module';
 
 @Module({
