+++ packages/twenty-server/src/workspace/messaging/repositories/message-participant/message-participant.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 
-import { CreateCompanyModule } from 'src/workspace/messaging/create-company/create-company.module';
-import { CreateContactModule } from 'src/workspace/messaging/create-contact/create-contact.module';
-import { MessageParticipantService } from 'src/workspace/messaging/message-participant/message-participant.service';
+import { CreateCompanyModule } from 'src/workspace/messaging/services/create-company/create-company.module';
+import { CreateContactModule } from 'src/workspace/messaging/services/create-contact/create-contact.module';
+import { MessageParticipantService } from 'src/workspace/messaging/repositories/message-participant/message-participant.service';
 import { WorkspaceDataSourceModule } from 'src/workspace/workspace-datasource/workspace-datasource.module';
 
 @Module({
