+++ packages/twenty-server/src/workspace/messaging/repositories/message-thread/message-thread.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 
-import { MessageChannelMessageAssociationModule } from 'src/workspace/messaging/message-channel-message-association/message-channel-message-assocation.module';
-import { MessageThreadService } from 'src/workspace/messaging/message-thread/message-thread.service';
+import { MessageChannelMessageAssociationModule } from 'src/workspace/messaging/repositories/message-channel-message-association/message-channel-message-assocation.module';
+import { MessageThreadService } from 'src/workspace/messaging/repositories/message-thread/message-thread.service';
 import { WorkspaceDataSourceModule } from 'src/workspace/workspace-datasource/workspace-datasource.module';
 
 @Module({
