+++ packages/twenty-server/src/workspace/messaging/repositories/workspace-member/workspace-member.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { WorkspaceMemberService } from 'src/workspace/messaging/workspace-member/workspace-member.service';
+import { WorkspaceMemberService } from 'src/workspace/messaging/repositories/workspace-member/workspace-member.service';
 import { WorkspaceDataSourceModule } from 'src/workspace/workspace-datasource/workspace-datasource.module';
 
 // TODO: Move outside of the messaging module
