+++ packages/twenty-server/src/workspace/messaging/services/create-contact/create-contact.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 
-import { CreateContactService } from 'src/workspace/messaging/create-contact/create-contact.service';
+import { CreateContactService } from 'src/workspace/messaging/services/create-contact/create-contact.service';
 
 @Module({
   imports: [],
