+++ packages/twenty-server/src/workspace/messaging/services/providers/messaging-providers.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 
 import { EnvironmentModule } from 'src/integrations/environment/environment.module';
-import { GmailClientProvider } from 'src/workspace/messaging/providers/gmail/gmail-client.provider';
+import { GmailClientProvider } from 'src/workspace/messaging/services/providers/gmail/gmail-client.provider';
 
 @Module({
   imports: [EnvironmentModule],
